Extract sync method and ignore full day events

+++ src/sync.py
@@ -17,32 +17,17 @@ def sync_to_nc(events: List[Event]) -> None:
         create_nc(ev)
 
 
-def sync_next_days_to_nc(amount: int) -> None:
+def sync_to(amount: int, target: str) -> None:
     today = datetime.datetime.today()
 
     for i in range(amount + 1):
         day = today + datetime.timedelta(days=i)
-        day_nc = get_day_nc(day, Calendar.FOR_NC_SYNC)
-        day_go = get_day_google(day, Calendar.FOR_NC_SYNC)
+        day_nc = get_day_nc(day, Calendar.FOR_GO_SYNC if target == 'go' else Calendar.FOR_NC_SYNC)
+        day_go = get_day_google(day, Calendar.FOR_GO_SYNC if target == 'go' else Calendar.FOR_NC_SYNC)
         equal = compare_day(day_nc, day_go)
 
-        print(equal)
         if not equal:
-            for ev in day_nc:
-                delete_event_nc(ev)
-            sync_to_nc(day_go)
-
-
-def sync_next_days_to_go(amount: int) -> None:
-    today = datetime.datetime.today()
-
-    for i in range(amount + 1):
-        day = today + datetime.timedelta(days=i)
-        day_nc = get_day_nc(day, Calendar.FOR_GO_SYNC)
-        day_go = get_day_google(day, Calendar.FOR_GO_SYNC)
-        equal = compare_day(day_nc, day_go)
-        print(equal)
-        if not equal:
-            for ev in day_go:
-                delete_event_go(ev)
-            sync_to_google(day_nc)
+            print(f'Day {day} not equal.\nNC: {day_nc}\nGO:{day_go}')
+            for ev in (day_go if target == 'go' else day_nc):
+                delete_event_go(ev) if target == 'go' else delete_event_nc(ev)
+            sync_to_google(day_nc) if target == 'go' else sync_to_nc(day_go)
